# docs/components/backend/analytics/openapi.json
{
  "components": {
    "schemas": {
...
      "ColumnListResponse": {
        "description": "Response envelope for `GET /v1/columns` and `GET /v1/columns/{table}`\n(`{ \"items\": [TableColumn] }`).\n\nDocs-only wrapper mirroring the inline `serde_json::json!` shape the\nhandlers emit — gives the column-list endpoints a real OpenAPI schema.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TableColumn"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "CreateRequest": {
        "additionalProperties": false,
        "description": "`POST /v1/admin/metric-thresholds` body — create a new threshold row.\n\n`tenant_id` / `id` / `locked_by` / `locked_at` / `created_at` /\n`updated_at` are NOT accepted from the body. `deny_unknown_fields`\nenforces that at the serde layer.\n\n`role_slug` / `team_id` use `Option<String>` — `None` is the canonical\nempty-string sentinel (DESIGN §3.7 + `infra/cache/catalog_cache.rs::cache_field`).",
        "properties": {
          "alert_bad": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "alert_trigger": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "good": {
            "format": "double",
            "type": "number"
          },
          "is_locked": {
            "type": "boolean"
          },
          "lock_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "metric_id": {
            "format": "uuid",
            "type": "string"
          },
          "role_slug": {
            "type": [
              "string",
              "null"
            ]
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "team_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "warn": {
            "format": "double",
            "type": "number"
          }
        },
        "required": [
          "metric_id",
          "scope",
          "good",
          "warn"
        ],
        "type": "object"
      },
...
      "ListResponse": {
        "description": "`GET /v1/admin/metric-thresholds` response envelope.\n\nWraps `items` in an object (instead of a bare array) so future\nadditions (pagination cursor, count, generated-at) are additive and\nnon-breaking. Mirrors the catalog read endpoint's envelope shape.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminMetricThresholdView"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem+json. `context` varies by error category.",
        "properties": {
          "context": {
            "type": "object"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "context"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read-only query service over predefined ClickHouse metrics. Admins define metrics (named SQL queries) in MariaDB; the frontend queries them by UUID with OData-style filtering. The API Gateway mounts this service at /api/analytics.",
    "title": "Analytics API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/admin/metric-thresholds": {
      "get": {
        "operationId": "analytics_api.admin.thresholds.list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            },
            "description": "List of metric thresholds"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List admin metric thresholds"
      },
      "post": {
        "operationId": "analytics_api.admin.thresholds.create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "description": "Metric threshold to create",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMetricThresholdView"
                }
              }
            },
            "description": "Created metric threshold"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create an admin metric threshold"
      }
    },
...
    "/v1/columns": {
      "get": {
        "operationId": "analytics_api.columns.list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColumnListResponse"
                }
              }
            },
            "description": "List of columns"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List queryable columns"
      }
    },
...
  }
}